fix dropdowns in dialogs going behind the keyboard (#4913)
I tried to find a solution without using a deprecated api to no avail.
But better a working app than perfect code I guess.

In the end this probably is a bug in the Material component library
https://github.com/material-components/material-components-android/issues/3307.

I also made the insets smaller which improves the dialogs a lot on small
screens.