feat: choose boost visibility (#3095) (#4944)
Closes #3095

---

Since I had to add `visibility` on the `onReblog` method, it is possible
that I have broken something. Also, I kept the old method signature
(which calls the new one), but it's possible that it is not needed.

~~I'm not 100% sure that unlisted boost works, as the visibility of the
boost is not shown anywehere.~~ I've confirmed that private
(followers-only) boosts are not visible on private browsing on the
browser.

EDIT: Confirmed visibility in devtools on web browser.

Screenshots:


![image](https://github.com/user-attachments/assets/f5bb8bcd-d13a-4263-8b2f-59f25621a6b1)
![image](https://github.com/user-attachments/assets/4e050e96-151d-45a5-9cd9-a7afd3a42c17)