improve WindowInsetsCallback to work when keyboard changes size (#4947)
re: https://chaos.social/@jannik/114059313885014801

Not sure why I didn't use `onPrepare` previously 🤔 . Makes more sense
this way.

The difference between `onStart` and `onPrepare` is that `onPrepare` is
called before the `OnApplyWindowInsetsListener` and the other after. So
with `onPrepare` it is easy to detect that an animation is starting. The
keyboard only seems to animate initially, not when it changes size
later, but the previous code always expected an animation.