Fix rtlsymmetry lint warnings (#5024)
This fixes #5001 

- I added some symmetry in views where I thought it is appropriate. 
- Then I added some 0 paddings in views where the warning was a false
positive.
- I also moved paddings to another view in the hierarchy in some places.
I think it makes the view more correct for the following textview to
have a start and endpadding instead of the view before that to have only
an end padding.

In the end I recreated the baseline file. 
It seems that two issues that I did not work on where fixed previously
that is why not only RtlSymmetry warnings are removed from the file.

---------

Co-authored-by: Janusz Leidgens <>