fix conversations jumping down on first load (#5006)
When opening the conversations tab with empty cache (i.e. after a frash
login), it would sometimes start with 6 year old messages for me, which
I find very annoying. The issue is that the loading footer is shown
first and the RecyclerView tries to keep it visible when new items are
added. This works around the issue. Note that it does not happen when
there are already some conversations in the cache.