fix undead MainActivity appearing after browser login and back press (#5056)
To reproduce, login an additional account via Browser login and then
press back. You will find yourself in a broken interface where the
displayed posts don't match the logged-in account. This does not happen
with Tusky login.
Seems like `finishAffinity()` does not work as expected when there is an
Activity of another app (the browser) on the stack.
Fix with the nuclear option (starting a new task). Also remove the
explode animation, as this was broken before (finish needs to be called
after startActivity for the animation to appear) and animations between
tasks don't work in newer Android versions anyway.