improve rendering quality when zooming into large images (#5068)
Glide downscales images to fit ImageViews, which makes perfect sense,
except here where the ImageView is zoomable.
Always loading the memory in full resolution of course uses more memory
but I think this is worth the additional quality. If it leds to
performance problems or crashes I will check if we can mitigate those in
another way.

Before / after:


![Screenshot_20250421_111204](https://github.com/user-attachments/assets/f00e9e54-d78b-40af-b2fc-0cfb83e588a6)

![Screenshot_20250421_110843](https://github.com/user-attachments/assets/b8319582-d149-4e47-ac08-7eb8a46be208)