show placeholders for null items in database backed recyclerviews (#5074)
When using the database as `PagingSource`, it can happen that there are
null items, e.g. when the user jumps to the top and the items are no
longer in memory and need to be re-queried. We have ignored this fact
until now, leading to subtle bugs where the adapter just shows a
completely empty `ViewHolder`, or worse, a recycled ViewHolder that has
not been updated and shows the wrong post. Usually these are only
visible for a split second but it can take longer in some cases e.g. on
slow devices.

Here is how the placeholders look:

<img
src="https://github.com/user-attachments/assets/58d3434f-916f-44a5-ad82-2a4a759e39d8"
width="320"/>


Note: I would prefer to turn this behavior with the null items off, but
we tried that once and it led to even worse bugs:
https://github.com/tuskyapp/Tusky/pull/4471