Ensure correct font weight for boost/fav count in all locales (#5105)
This never worked as intended when the font weight was defined inside
the resources. Seems like Weblate is not dealing with that correctly.
If there is `<b>%1$s</b>` used inside the resource, it gets stripped by
`getQuantityString`. To work around that, it needs to be HTML escaped
`&lt;b>%1$s&lt;/b>` and then passed to `fromHtml`.

Since for some reason not even the source strings got the
`&lt;b>%1$s&lt;/b>` correct everywhere, I'm changing it to a completely
code based version that will always be right no matter what is defined
in the resource.